# Improve performance by adding hit_counts table