--- a/cron/hit_stat.go
+++ b/cron/hit_stat.go
@@ -10,6 +10,7 @@ import (
 	"strconv"
 
 	"zgo.at/goatcounter"
+	"zgo.at/goatcounter/cfg"
 	"zgo.at/goatcounter/errors"
 	"zgo.at/utils/jsonutil"
 	"zgo.at/zdb"
@@ -31,7 +32,10 @@ func updateHitStats(ctx context.Context, hits []goatcounter.Hit) error {
 		type gt struct {
 			count       []int
 			countUnique []int
+			total       int
+			totalUnique int
 			day         string
+			hour        string
 			event       zdb.Bool
 			path        string
 			title       string
@@ -43,10 +47,12 @@ func updateHitStats(ctx context.Context, hits []goatcounter.Hit) error {
 			}
 
 			day := h.CreatedAt.Format("2006-01-02")
+			dayHour := h.CreatedAt.Format("2006-01-02 15:00:00")
 			k := fmt.Sprintf("%s%s%t", day, h.Path, h.Event)
 			v := grouped[k]
 			if len(v.count) == 0 {
 				v.day = day
+				v.hour = dayHour
 				v.path = h.Path
 				v.event = h.Event
 				var err error
@@ -54,6 +60,12 @@ func updateHitStats(ctx context.Context, hits []goatcounter.Hit) error {
 					h.Site, day, v.path, v.event)
 				if err != nil {
 					return err
+				}
+				for _, i := range v.count {
+					v.total += i
+				}
+				for _, i := range v.countUnique {
+					v.totalUnique += i
 				}
 			}
 
@@ -63,7 +75,9 @@ func updateHitStats(ctx context.Context, hits []goatcounter.Hit) error {
 
 			hour, _ := strconv.ParseInt(h.CreatedAt.Format("15"), 10, 8)
 			v.count[hour] += 1
+			v.total += 1
 			if h.FirstVisit {
+				v.totalUnique += 1
 				v.countUnique[hour] += 1
 			}
 			grouped[k] = v
@@ -76,8 +90,27 @@ func updateHitStats(ctx context.Context, hits []goatcounter.Hit) error {
 			ins.Values(siteID, v.day, v.path, v.event, v.title,
 				jsonutil.MustMarshal(v.count),
 				jsonutil.MustMarshal(v.countUnique))
+
+			var err error
+			if cfg.PgSQL {
+				_, err = zdb.MustGet(ctx).ExecContext(ctx, `insert into hit_counts
+				(site, path, title, event, hour, total, total_unique) values ($1, $2, $3, $4, $5, $6, $7)
+				on conflict on constraint "hit_counts#site#path#hour#event" do
+					update set total=$8, total_unique=$9`,
+					siteID, v.path, v.title, v.event, v.hour, v.total, v.totalUnique,
+					v.total, v.totalUnique)
+			} else {
+				// SQLite has "on conflict replace" on the unique constraint to
+				// do the same.
+				_, err = zdb.MustGet(ctx).ExecContext(ctx, `insert into hit_counts
+					(site, path, title, event, hour, total, total_unique) values ($1, $2, $3, $4, $5, $6, $7)`,
+					siteID, v.path, v.title, v.event, v.hour, v.total, v.totalUnique)
+			}
+			if err != nil {
+				return errors.Wrap(err, "updateHitStats hit_counts")
+			}
 		}
-		return errors.Wrap(ins.Finish(), "updateHitStats")
+		return errors.Wrap(ins.Finish(), "updateHitStats hit_stats")
 	})
 }