--- a/go.mod
+++ b/go.mod
@@ -19,8 +19,8 @@ require (
 	zgo.at/isbot v0.0.0-20200427181629-1d4940c02e50
 	zgo.at/tz v0.0.0-20200428173314-f4a46a753a7e
 	zgo.at/utils v0.0.0-20200514044306-bf7c1ff8aa0c
-	zgo.at/zdb v0.0.0-20200514044421-e7d35d5f79f1
-	zgo.at/zhttp v0.0.0-20200514040227-c90f02522ddf
+	zgo.at/zdb v0.0.0-20200518092829-2353fffa61c0
+	zgo.at/zhttp v0.0.0-20200518165501-40ecbc2139a8
 	zgo.at/zlog v0.0.0-20200427184116-582329a88e79
 	zgo.at/zpack v1.0.1
 	zgo.at/zstripe v1.0.0