--- a/pack/pack.go
+++ b/pack/pack.go
@@ -422,6 +422,36 @@ commit;
 	insert into version values ('2020-05-17-1-rm-user-name');
 commit;
 `),
+	"db/migrate/pgsql/2020-05-18-1-domain-count.sql": []byte(`begin;
+	create table hit_counts (
+		site          int        not null check(site>0),
+		path          varchar    not null,
+		title         varchar    not null,
+		event         integer    not null default 0,
+		hour          timestamp  not null,
+		total         int        not null,
+		total_unique  int        not null,
+
+		constraint "hit_counts#site#path#hour#event" unique(site, path, hour, event)
+	);
+	create index "hit_counts#site#hour#event" on hit_counts(site, hour, event);
+
+	insert into hit_counts (site, path, title, event, hour, total, total_unique)
+		select
+				site,
+				max(path),
+				max(title) as title,
+				event,
+				(substring(created_at::varchar, 0, 14) || ':00:00')::timestamp as hour,
+				count(*),
+				sum(first_visit)
+		from hits
+		where bot=0
+		group by site, lower(path), event, hour;
+
+	insert into version values ('2020-05-18-1-domain-count');
+commit;
+`),
 }
 
 var MigrationsSQLite = map[string][]byte{
@@ -915,6 +945,36 @@ commit;
 	create unique index "sites#code" on sites(lower(code));
 
 	insert into version values ('2020-05-17-1-rm-user-name');
+commit;
+`),
+	"db/migrate/sqlite/2020-05-18-1-domain-count.sql": []byte(`begin;
+	create table hit_counts (
+		site          int        not null check(site>0),
+		path          varchar    not null,
+		title         varchar    not null,
+		event         integer    not null default 0,
+		hour          timestamp  not null check(hour = strftime('%Y-%m-%d %H:%M:%S', hour)),
+		total         int        not null,
+		total_unique  int        not null,
+
+		constraint "hit_counts#site#path#hour#event" unique(site, path, hour, event) on conflict replace
+	);
+	create index "hit_counts#site#hour#event" on hit_counts(site, hour, event);
+
+	insert into hit_counts (site, path, title, event, hour, total, total_unique)
+		select
+				site,
+				max(path),
+				max(title) as title,
+				event,
+				(substr(created_at, 0, 14) || ':00:00') as hour,
+				count(*),
+				sum(first_visit)
+		from hits
+		where bot=0
+		group by site, lower(path), event, hour;
+
+	insert into version values ('2020-05-18-1-domain-count');
 commit;
 `),
 }
@@ -15103,12 +15163,13 @@ Martin
 <h2>Sites</h2>
 <p>All sites with at least 1,000 hits in the last 30 days; the counts for the
 parent site includes the child sites.</p>
-<table>
+<table style="max-width: none">
 	<tr>
 		<th>ID</th>
-		<th><a href="?order=count"># hits</a></th>
+		<th><a href="?order=last_month">hits/month</a></th>
+		<th><a href="?order=total">total hits</a></th>
 		<th>Code</th>
-		<th>Name</th>
+		<th>Domain</th>
 		<th>User</th>
 		<th>Plan</th>
 		<th><a href="?order=created_at">Created at</a></th>
@@ -15116,7 +15177,8 @@ parent site includes the child sites.</p>
 	{{range $s := .Stats}}
 		<tr id="{{$s.ID}}" class="plan-{{$s.Plan}}">
 			<td><a href="/admin/{{$s.ID}}">{{$s.ID}}</a></td>
-			<td>{{nformat $s.Count $.Site}}</td>
+			<td>{{nformat $s.LastMonth $.Site}}</td>
+			<td>{{nformat $s.Total $.Site}}</td>
 			<td>
 				{{if $s.Public}}
 					<a href="https://{{$s.Code}}.{{$.Domain}}">{{$s.Code}}</a>
@@ -15124,11 +15186,23 @@ parent site includes the child sites.</p>
 					{{$s.Code}}
 				{{end}}
 			</td>
-			<td>{{if $s.LinkDomain}} – {{$s.LinkDomain}}{{end}}</td>
-			<td>"{{$s.User}}" &lt;{{$s.Email}}&gt;</td>
+			<td>{{$s.LinkDomain}}</td>
+			<td>{{$s.Email}}</td>
 			<td>
-				{{$s.Plan}}
-				{{if eq $s.Plan "child"}}(<a href="#{{$s.Parent}}">{{$s.Parent}}</a>){{end}}
+				{{if $s.Stripe}}
+					{{$s.Plan}}
+					{{if has_prefix $s.Stripe "cus_github"}}
+					<a href="https://github.com/{{substr $s.Stripe 11 -1}}">GitHub</a>
+					{{else if not (has_prefix $s.Stripe "cus_free_")}}
+						<a href="https://dashboard.stripe.com/customers/{{$s.Stripe}}">Stripe</a>
+					{{end}}
+				{{else}}
+					{{if eq $s.Plan "child"}}
+						child of <a href="#{{$s.Parent}}">{{$s.Parent}}</a>
+					{{else}}
+						free
+					{{end}}
+				{{end}}
 			</td>
 			<td>{{tformat $.Site $s.CreatedAt ""}}</td>
 		</tr>