--- a/tpl/backend_admin.gohtml
+++ b/tpl/backend_admin.gohtml
@@ -12,12 +12,13 @@
 <h2>Sites</h2>
 <p>All sites with at least 1,000 hits in the last 30 days; the counts for the
 parent site includes the child sites.</p>
-<table>
+<table style="max-width: none">
 	<tr>
 		<th>ID</th>
-		<th><a href="?order=count"># hits</a></th>
+		<th><a href="?order=last_month">hits/month</a></th>
+		<th><a href="?order=total">total hits</a></th>
 		<th>Code</th>
-		<th>Name</th>
+		<th>Domain</th>
 		<th>User</th>
 		<th>Plan</th>
 		<th><a href="?order=created_at">Created at</a></th>
@@ -25,7 +26,8 @@ parent site includes the child sites.</p>
 	{{range $s := .Stats}}
 		<tr id="{{$s.ID}}" class="plan-{{$s.Plan}}">
 			<td><a href="/admin/{{$s.ID}}">{{$s.ID}}</a></td>
-			<td>{{nformat $s.Count $.Site}}</td>
+			<td>{{nformat $s.LastMonth $.Site}}</td>
+			<td>{{nformat $s.Total $.Site}}</td>
 			<td>
 				{{if $s.Public}}
 					<a href="https://{{$s.Code}}.{{$.Domain}}">{{$s.Code}}</a>
@@ -33,11 +35,23 @@ parent site includes the child sites.</p>
 					{{$s.Code}}
 				{{end}}
 			</td>
-			<td>{{if $s.LinkDomain}} – {{$s.LinkDomain}}{{end}}</td>
-			<td>"{{$s.User}}" &lt;{{$s.Email}}&gt;</td>
+			<td>{{$s.LinkDomain}}</td>
+			<td>{{$s.Email}}</td>
 			<td>
-				{{$s.Plan}}
-				{{if eq $s.Plan "child"}}(<a href="#{{$s.Parent}}">{{$s.Parent}}</a>){{end}}
+				{{if $s.Stripe}}
+					{{$s.Plan}}
+					{{if has_prefix $s.Stripe "cus_github"}}
+					<a href="https://github.com/{{substr $s.Stripe 11 -1}}">GitHub</a>
+					{{else if not (has_prefix $s.Stripe "cus_free_")}}
+						<a href="https://dashboard.stripe.com/customers/{{$s.Stripe}}">Stripe</a>
+					{{end}}
+				{{else}}
+					{{if eq $s.Plan "child"}}
+						child of <a href="#{{$s.Parent}}">{{$s.Parent}}</a>
+					{{else}}
+						free
+					{{end}}
+				{{end}}
 			</td>
 			<td>{{tformat $.Site $s.CreatedAt ""}}</td>
 		</tr>