This adds a new hit_counts table with intermediate counts. This improves
performance by quite a bit.

This is really how hit_stats should have been, instead of aggregating all the
daily stats in one stats row (which is hard to query with TZ support). I might
migrate to this in the future and drop hit_stats, but for now this is fine.

Before on my site when selecting "last year":

	dashboard        40ms  browsers.List
	dashboard        50ms  sizeStat.ListSizes
	dashboard        78ms  locStat.List
	dashboard       111ms  topRefs.List
	HitStats.List  4900ms  select hits
	HitStats.List   134ms  select hits_stats
	HitStats.List    59ms  add hit_stats
	HitStats.List     2ms  fill blanks
	HitStats.List     1ms  tz
	HitStats.List     0ms  add totals
	dashboard      5098ms  get data
	dashboard        19ms  zhttp.Template

After:

	dashboard        27ms  browsers.List
	dashboard        21ms  sizeStat.ListSizes
	dashboard        39ms  locStat.List
	HitStats.List   115ms  select hits
	dashboard        37ms  topRefs.List
	HitStats.List    98ms  select hits_stats
	HitStats.List    62ms  add hit_stats
	HitStats.List     1ms  fill blanks
	HitStats.List     1ms  tz
	HitStats.List     0ms  add totals
	dashboard       280ms  get data
	dashboard        10ms  zhttp.Template

Largest site selecting "last year":

	dashboard        15ms  browsers.List
	dashboard        17ms  sizeStat.ListSizes
	dashboard        35ms  locStat.List
	dashboard        18ms  topRefs.List
	HitStats.List  6861ms  select hits
	HitStats.List    25ms  select hits_stats
	HitStats.List    19ms  add hit_stats
	HitStats.List     2ms  fill blanks
	HitStats.List     1ms  tz
	HitStats.List     1ms  add totals
	dashboard      7182ms  get data
	dashboard        12ms  zhttp.Template

After:

	dashboard         9ms  browsers.List
	HitStats.List    13ms  select hits
	dashboard         9ms  sizeStat.ListSizes
	HitStats.List     8ms  select hits_stats
	dashboard         9ms  locStat.List
	dashboard         5ms  topRefs.List
	HitStats.List    13ms  add hit_stats
	HitStats.List     2ms  fill blanks
	HitStats.List     9ms  tz
	HitStats.List     1ms  add totals
	dashboard        49ms  get data
	dashboard        16ms  zhttp.Template

My site is slower not because it has more hits, but because it has a larger
number of entries because it's older and has more paths with consistent hits:

	(1)=# select count(*), sum(total) from hit_counts where site=1;
	count  │  sum
	───────┼────────
	80845  │ 443918

	(1)=# select count(*), sum(total) from hit_counts where site=133;
	count  │   sum
	───────┼─────────
	4140   │ 2522877